Refactor AiTemporaryStorage usage
